Percentile height or weight for age for children

This is the underlying function, the exposed functions are pct_weight_for_age() and pct_height_for_age() Based on tables from WHO: http://www.who.int/childgrowth/standards/height_for_age/en/

Usage

pct_for_age_generic(
  age = NULL,
  value = NULL,
  sex = NULL,
  variable = "weight",
  ...
)

Arguments

age

age in years

value

height in kg. Optional, if specified, will calculate closest percentile and return in list as `percentile`

sex

either `male` or `female`

variable

weight or height?

...

parameters passed to `read_who_table()`